split off nightmare dependencies, bump version for interim build.
