+++ app/package.json
@@ -1,7 +1,7 @@
 {
   "name": "balancebeam",
   "productName": "BalanceBeam",
-  "version": "1.2.2",
+  "version": "1.3.0",
   "description": "A Desktop application for cash-flow management",
   "main": "./main.prod.js",
   "author": {
